# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Universal AI Coding Proxy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /proxy/ask:
    post:
      operationId: askAi
      tags: [proxy]
      summary: Send a prompt to an AI
      description: Routes a coding prompt to the selected AI service via browser automation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskAiBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiResponse"
        "503":
          description: AI service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /proxy/execute:
    post:
      operationId: executeCode
      tags: [proxy]
      summary: Execute generated code
      description: Runs a code snippet in a sandboxed subprocess
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteCodeBody"
      responses:
        "200":
          description: Execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionResult"
...
  /proxy/set-model:
    post:
      operationId: setModel
      tags: [proxy]
      summary: Set the active model for an AI service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                aiId:
                  type: string
                modelId:
                  type: string
              required:
                - aiId
                - modelId
      responses:
        "200":
          description: Model set successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  aiId:
                    type: string
                  modelId:
                    type: string
...
  /sessions/create:
    post:
      operationId: createSession
      tags: [sessions]
      summary: Create a new browser session for an AI
      description: Launches a visible browser window for manual login, then saves the session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionBody"
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
  /sessions/{aiId}/delete:
    delete:
      operationId: deleteSession
      tags: [sessions]
      summary: Delete a saved session
      parameters:
        - name: aiId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
...
  /history/{aiId}:
    get:
      operationId: getHistory
      tags: [history]
      summary: Get conversation history for a specific AI
      parameters:
        - name: aiId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Conversation history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationHistory"

    delete:
      operationId: clearHistory
      tags: [history]
      summary: Clear conversation history for a specific AI
      parameters:
        - name: aiId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: History cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
...
  /files/tree:
    get:
      operationId: getFileTree
      tags: [files]
      summary: Get directory tree
      parameters:
        - name: path
          in: query
          schema:
            type: string
            default: "."
          description: Root path to list (relative to workspace root)
        - name: depth
          in: query
          schema:
            type: integer
            default: 3
          description: Max depth to traverse
      responses:
        "200":
          description: Directory tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileTreeResponse"
  /files/read:
    get:
      operationId: readFile
      tags: [files]
      summary: Read a file's contents
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
          description: File path (relative to workspace root)
      responses:
        "200":
          description: File contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileReadResponse"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /files/write:
    post:
      operationId: writeFile
      tags: [files]
      summary: Write content to a file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileWriteBody"
      responses:
        "200":
          description: File written
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileWriteResponse"
  /files/create:
    post:
      operationId: createFile
      tags: [files]
      summary: Create a new file or directory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileCreateBody"
      responses:
        "200":
          description: File or directory created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileWriteResponse"
  /files/delete:
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file or directory
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileWriteResponse"
  /files/rename:
    post:
      operationId: renameFile
      tags: [files]
      summary: Rename or move a file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileRenameBody"
      responses:
        "200":
          description: File renamed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileWriteResponse"
  /terminal/exec:
    post:
      operationId: terminalExec
      tags: [proxy]
      summary: Execute a shell command
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TerminalExecBody"
      responses:
        "200":
          description: Command output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerminalExecResponse"
...
  /agent/run:
    post:
      operationId: runAgent
      tags: [proxy]
      summary: Run an autonomous coding agent task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRunBody"
      responses:
        "200":
          description: Agent task trace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRunResponse"
        "503":
          description: AI unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /proxy/ask-with-context:
    post:
      operationId: askAiWithContext
      tags: [proxy]
      summary: Send a prompt to AI with file context attached
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskWithContextBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiResponse"
        "503":
          description: AI unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    AskAiBody:
      type: object
      properties:
        aiId:
          type: string
          description: ID of the AI to use (e.g., "chatgpt", "grok")
        prompt:
          type: string
          description: The coding prompt or question
        conversationId:
          type: string
          nullable: true
          description: Optional conversation ID to continue a session
        fallback:
          type: boolean
          default: true
          description: Whether to automatically fallback to another AI on failure
      required:
        - aiId
        - prompt
    AiResponse:
      type: object
      properties:
        success:
          type: boolean
        aiId:
          type: string
        response:
          type: string
        conversationId:
          type: string
        elapsedMs:
          type: number
        fallbackUsed:
          type: boolean
        error:
          type: string
          nullable: true
      required:
        - success
        - aiId
        - response
        - conversationId
        - elapsedMs
        - fallbackUsed
    ExecuteCodeBody:
      type: object
      properties:
        code:
          type: string
          description: Code to execute
        language:
          type: string
          default: "python"
          description: Programming language (python, javascript, bash)
        timeout:
          type: integer
          default: 30
          description: Execution timeout in seconds
      required:
        - code
    ExecutionResult:
      type: object
      properties:
        success:
          type: boolean
        stdout:
          type: string
        stderr:
          type: string
        exitCode:
          type: integer
        elapsedMs:
          type: number
      required:
        - success
        - stdout
        - stderr
        - exitCode
        - elapsedMs
...
    CreateSessionBody:
      type: object
      properties:
        aiId:
          type: string
          description: ID of the AI to create a session for
      required:
        - aiId
    SessionResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        aiId:
          type: string
      required:
        - success
        - message
...
    ConversationHistory:
      type: object
      properties:
        aiId:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
      required:
        - aiId
        - messages
    Message:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        timestamp:
          type: string
          format: date-time
        conversationId:
          type: string
      required:
        - id
        - role
        - content
        - timestamp
        - conversationId
...
    FileNode:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory]
        size:
          type: integer
          nullable: true
        extension:
          type: string
          nullable: true
        children:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/FileNode"
      required:
        - name
        - path
        - type
    FileTreeResponse:
      type: object
      properties:
        tree:
          $ref: "#/components/schemas/FileNode"
        rootPath:
          type: string
      required:
        - tree
        - rootPath
    FileReadResponse:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
        size:
          type: integer
        extension:
          type: string
          nullable: true
        lineCount:
          type: integer
        isBinary:
          type: boolean
      required:
        - path
        - content
        - size
        - lineCount
        - isBinary
    FileWriteBody:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
      required:
        - path
        - content
    FileCreateBody:
      type: object
      properties:
        path:
          type: string
        type:
          type: string
          enum: [file, directory]
          default: file
        content:
          type: string
          nullable: true
      required:
        - path
    FileRenameBody:
      type: object
      properties:
        oldPath:
          type: string
        newPath:
          type: string
      required:
        - oldPath
        - newPath
    FileWriteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        path:
          type: string
      required:
        - success
        - message
        - path
    AskWithContextBody:
      type: object
      properties:
        aiId:
          type: string
        prompt:
          type: string
        files:
          type: array
          items:
            $ref: "#/components/schemas/FileContext"
        action:
          type: string
          nullable: true
          description: Quick action like "fix", "explain", "test", "refactor", "suggest"
        conversationId:
          type: string
          nullable: true
        fallback:
          type: boolean
          default: true
      required:
        - aiId
        - prompt
    FileContext:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
        language:
          type: string
          nullable: true
      required:
        - path
        - content
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    TerminalExecBody:
      type: object
      properties:
        command:
          type: string
          description: Shell command to execute
        cwd:
          type: string
          nullable: true
          description: Working directory (defaults to workspace root)
        timeout:
          type: integer
          default: 60
          description: Timeout in seconds
      required:
        - command
    TerminalExecResponse:
      type: object
      properties:
        stdout:
          type: string
        stderr:
          type: string
        exitCode:
          type: integer
        elapsedMs:
          type: number
        cwd:
          type: string
      required:
        - stdout
        - stderr
        - exitCode
        - elapsedMs
        - cwd
...
    AgentRunBody:
      type: object
      properties:
        aiId:
          type: string
          description: AI to use for the agent
        task:
          type: string
          description: Task description — what to build or do
        workingDir:
          type: string
          nullable: true
          description: Working directory for the agent
        maxSteps:
          type: integer
          default: 20
          description: Maximum number of tool calls
      required:
        - aiId
        - task
    AgentStep:
      type: object
      properties:
        step:
          type: integer
        type:
          type: string
          enum: [thought, tool, error]
        content:
          type: string
          nullable: true
        tool:
          type: string
          nullable: true
        params:
          type: object
          nullable: true
          additionalProperties: true
        result:
          type: string
          nullable: true
        elapsedMs:
          type: number
      required:
        - step
        - type
    AgentResult:
      type: object
      properties:
        success:
          type: boolean
        summary:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
        totalElapsedMs:
          type: number
          nullable: true
      required:
        - success
    AgentRunResponse:
      type: object
      properties:
        running:
          type: boolean
        task:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/AgentStep"
        result:
          $ref: "#/components/schemas/AgentResult"
          nullable: true
      required:
        - running
        - task
        - steps
...